requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
fonttools==4.33.3
frozenlist==1.3.0
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
langdetect==1.0.9
matplotlib==3.5.2
multidict==6.0.2
nltk==3.7
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pycountry==22.3.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.0
tweepy==4.9.0
urllib3==1.26.9
wordcloud==1.8.1
yarl==1.7.2
